Add testing for outer level v1beta1 provisioning controller package
