+++ pkg/apis/apis.go
@@ -38,7 +38,7 @@ var (
 	Settings    = []settings.Injectable{&settings.Settings{}}
 )
 
-//go:generate controller-gen crd object:headerFile="../../hack/boilerplate.go.txt" paths="./..." output:crd:artifacts:config=crds
+//go:generate controller-gen crd:generateEmbeddedObjectMeta=true object:headerFile="../../hack/boilerplate.go.txt" paths="./..." output:crd:artifacts:config=crds
 var (
 	//go:embed crds/karpenter.sh_provisioners.yaml
 	ProvisionerCRD []byte
