+++ pkg/apis/crds/karpenter.sh_nodepools.yaml
@@ -100,6 +100,23 @@ spec:
                   specifies.
                 properties:
                   metadata:
+                    properties:
+                      annotations:
+                        additionalProperties:
+                          type: string
+                        type: object
+                      finalizers:
+                        items:
+                          type: string
+                        type: array
+                      labels:
+                        additionalProperties:
+                          type: string
+                        type: object
+                      name:
+                        type: string
+                      namespace:
+                        type: string
                     type: object
                   spec:
                     description: NodeClaimSpec describes the desired state of the
