+++ pkg/apis/v1beta1/nodeclaim.go
@@ -131,10 +131,6 @@ type NodeClassReference struct {
 	// API version of the referent
 	// +optional
 	APIVersion string `json:"apiVersion,omitempty"`
-	// IsNodeTemplate tells Karpenter whether the in-memory representation of this object
-	// is actually referring to a NodeTemplate object. This value is not actually part of the v1beta1 public-facing API
-	// TODO @joinnis: Remove this field when v1alpha5 is unsupported in a future version of Karpenter
-	IsNodeTemplate bool `json:"-"`
 }
 
 // +kubebuilder:object:generate=false
